switch to float

+++ coral_ordinal/metrics.py
@@ -7,7 +7,7 @@ def MeanAbsoluteErrorLabels(y_true, y_pred):
   cum_probs = tf.map_fn(tf.math.sigmoid, y_pred)
   
   # Calculate the labels using the style of Cao et al.
-  labels_v2 = tf.reduce_sum(tf.map_fn(lambda x: tf.cast(x > 0.5, tf.int16), cum_probs), axis = 1)
+  labels_v2 = tf.reduce_sum(tf.map_fn(lambda x: tf.cast(x > 0.5, tf.float32), cum_probs), axis = 1)
   
   labels_v2 = tf.cast(labels_v2, y_true.dtype)
